requirements.txt
asgiref==3.5.2
Django==4.1.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
patsy==0.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
